GPII-1033: Starting to add preferences selection
Added a selection tree subcomponent to handle the preferences selection. The model from the selection tree is passed to the server as serialized JSON attached to a hidden input in the form.

There are a couple of styling issues. 1) the indeterminate checkboxes are missing the inner box 2) there is a problem caused by the line-heights set from Foundation that cause the toggles to be the wrong size and misaligned.

Further work will involve fetching the requestedPrefs via ajax, and pulling in the template for the selection tree.